Check AdminOperations CleanUp AdminOperations

+++ src/Database/CourseDatabase.java
@@ -31,13 +31,12 @@ public class CourseDatabase implements CourseDBOperations{
     }
 
     public void addCourse(Course course){
-
         this.courses.add(course);
     }
 
     public void deleteCourse(String courseId,String userId){
         for(Course course : this.courses){
-            if((userId.contains("adm") || course.getCreatorId().equals(userId)) && course.getCourseId().equals(courseId)){
+            if((userId.contains("Adm") || course.getCreatorId().equals(userId)) && course.getCourseId().equals(courseId)){
                 this.courses.remove(course);
                 break;
             }
@@ -70,14 +69,14 @@ public class CourseDatabase implements CourseDBOperations{
 
     public void addCourseCategory(String category, String courseId, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        if((userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId)) {
+        if(courseIndex!=-1 && (userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId)) {
             this.courses.get(courseIndex).addCourseCategory(category);
         }
     }
 
     public void removeCourseCategory(String category, String courseId, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        if((userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId)) {
+        if(courseIndex!=-1 && (userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId)) {
             this.courses.get(courseIndex).removeCourseCategory(category);
         }
     }
@@ -86,14 +85,15 @@ public class CourseDatabase implements CourseDBOperations{
 
     public ArrayList<Comment> getCourseComments(String courseId, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        if(this.courses.get(courseIndex).getCreatorId().equals(userId))
+        if(courseIndex!=-1 && this.courses.get(courseIndex).getCreatorId().equals(userId))
             return new ArrayList<>(this.courses.get(courseIndex).getComments());
         return new ArrayList<>();
     }
 
     public void addComment(String comment, String courseId, String userId){
         int courseIndex = getCourseIndex(courseId);
-        this.courses.get(courseIndex).addComment(comment,userId);
+        if(courseIndex!=-1)
+            this.courses.get(courseIndex).addComment(comment,userId);
     }
 
     //*************************************************************
@@ -123,45 +123,46 @@ public class CourseDatabase implements CourseDBOperations{
     //*************************************************************
     public void rateCourse(String courseId, int rating, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        this.courses.get(courseIndex).updateRating(rating,userId);
+        if(courseIndex!=-1)
+            this.courses.get(courseIndex).updateRating(rating,userId);
     }
 
     public void changeCourseName(String newCourseName,String courseId, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        if((userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId)) {
+        if(courseIndex!=-1 && (userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId)) {
             this.courses.get(courseIndex).changeCourseName(newCourseName);
         }
     }
 
     public void changeCoursePrice(int newPrice, String courseId, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        if((userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId)){
+        if(courseIndex!=-1 && (userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId)){
             this.courses.get(courseIndex).changePrice(newPrice);
         }
     }
 
     public void addCourseContent(String courseId, Chapter courseChapter, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        if((userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId))
+        if(courseIndex!=-1 && (userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId))
             this.courses.get(courseIndex).addContent(courseChapter);
     }
 
     public void deleteCourseContent(String courseId, int contentIndex, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        if((userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId))
+        if(courseIndex!=-1 && (userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId))
             this.courses.get(courseIndex).deleteContent(contentIndex);
 
     }
 
     public void changeCourseChapterName(String newChapterName, String courseId, int chapterIndex, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        if((userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId))
+        if(courseIndex!=-1 && (userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId))
             this.courses.get(courseIndex).changeChapterName(newChapterName,chapterIndex);
     }
 
     public void changeCourseLesson(String newLesson, String courseId, int chapterIndex, String userId) {
         int courseIndex = getCourseIndex(courseId);
-        if((userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId))
+        if(courseIndex!=-1 && (userId.contains("Adm") && courseId.contains("ZCourse")) || this.courses.get(courseIndex).getCreatorId().equals(userId))
             this.courses.get(courseIndex).changeLesson(newLesson,chapterIndex);
     }
